backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
python-dotenv==1.2.1
sqlparse==0.5.5
django-filter==25.2
pytest==9.0.2
pytest-django==4.12.0
stripe==11.5.0
